+++ src/Domain/Model/Subscription/Subscription.php
@@ -14,6 +14,7 @@ use Doctrine\ORM\Proxy\Proxy;
 use JMS\Serializer\Annotation\ExclusionPolicy;
 use JMS\Serializer\Annotation\Expose;
 use PhpList\PhpList4\Domain\Model\Interfaces\CreationDate;
+use PhpList\PhpList4\Domain\Model\Interfaces\DomainModel;
 use PhpList\PhpList4\Domain\Model\Interfaces\ModificationDate;
 use PhpList\PhpList4\Domain\Model\Messaging\SubscriberList;
 use PhpList\PhpList4\Domain\Model\Traits\CreationDateTrait;
@@ -30,7 +31,7 @@ use PhpList\PhpList4\Domain\Model\Traits\ModificationDateTrait;
  *
  * @author Oliver Klee <oliver@phplist.com>
  */
-class Subscription implements CreationDate, ModificationDate
+class Subscription implements DomainModel, CreationDate, ModificationDate
 {
     use CreationDateTrait;
     use ModificationDateTrait;