+++ src/Domain/Repository/AbstractRepository.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace PhpList\PhpList4\Domain\Repository;
 
 use Doctrine\ORM\EntityRepository;
-use PhpList\PhpList4\Domain\Model\Interfaces\Identity;
+use PhpList\PhpList4\Domain\Model\Interfaces\DomainModel;
 
 /**
  * Base class for repositories.
@@ -14,18 +14,34 @@ use PhpList\PhpList4\Domain\Model\Interfaces\Identity;
 abstract class AbstractRepository extends EntityRepository
 {
     /**
-     * Persists and $model and flushes the entity manager change list.
+     * Persists $model and flushes the entity manager change list.
      *
      * This method allows controllers to not depend on the entity manager, but only on the repositories instead,
      * following the Law of Demeter.
      *
-     * @param Identity $model
+     * @param DomainModel $model
      *
      * @return void
      */
-    public function save(Identity $model)
+    public function save(DomainModel $model)
     {
         $this->getEntityManager()->persist($model);
+        $this->getEntityManager()->flush();
+    }
+
+    /**
+     * Removes $model and flushes the entity manager change list.
+     *
+     * This method allows controllers to not depend on the entity manager, but only on the repositories instead,
+     * following the Law of Demeter.
+     *
+     * @param DomainModel $model
+     *
+     * @return void
+     */
+    public function remove(DomainModel $model)
+    {
+        $this->getEntityManager()->remove($model);
         $this->getEntityManager()->flush();
     }
 }