+++ tests/Integration/Domain/Repository/Identity/AdministratorTokenRepositoryTest.php
@@ -258,4 +258,23 @@ class AdministratorTokenRepositoryTest extends TestCase
 
         static::assertSame($model, $this->subject->find($model->getId()));
     }
+
+    /**
+     * @test
+     */
+    public function removeRemovesModel()
+    {
+        $this->getDataSet()->addTable(static::TABLE_NAME, __DIR__ . '/../Fixtures/DetachedAdministratorTokens.csv');
+        $this->applyDatabaseChanges();
+
+        /** @var AdministratorToken[] $allModels */
+        $allModels = $this->subject->findAll();
+        $numberOfModelsBeforeRemove = count($allModels);
+        $firstModel = $allModels[0];
+
+        $this->subject->remove($firstModel);
+
+        $numberOfModelsAfterRemove = count($this->subject->findAll());
+        static::assertSame(1, $numberOfModelsBeforeRemove - $numberOfModelsAfterRemove);
+    }
 }