+++ tests/Integration/Domain/Repository/Subscription/SubscriberRepositoryTest.php
@@ -293,6 +293,25 @@ class SubscriberRepositoryTest extends TestCase
 
         $newNumberOfSubscriptions = count($this->subscriptionRepository->findAll());
         $numberOfRemovedSubscriptions = $initialNumberOfSubscriptions - $newNumberOfSubscriptions;
-        static::assertSame($numberOfAssociatedSubscriptions, $newNumberOfSubscriptions);
+        static::assertSame($numberOfAssociatedSubscriptions, $numberOfRemovedSubscriptions);
+    }
+
+    /**
+     * @test
+     */
+    public function removeRemovesModel()
+    {
+        $this->getDataSet()->addTable(static::TABLE_NAME, __DIR__ . '/../Fixtures/Subscriber.csv');
+        $this->applyDatabaseChanges();
+
+        /** @var Subscriber[] $allModels */
+        $allModels = $this->subject->findAll();
+        $numberOfModelsBeforeRemove = count($allModels);
+        $firstModel = $allModels[0];
+
+        $this->subject->remove($firstModel);
+
+        $numberOfModelsAfterRemove = count($this->subject->findAll());
+        static::assertSame(1, $numberOfModelsBeforeRemove - $numberOfModelsAfterRemove);
     }
 }