+++ tests/Support/Interfaces/TestMarker.php
@@ -5,7 +5,7 @@ namespace PhpList\PhpList4\Tests\Support\Interfaces;
 
 /**
  * The presence of this interface means that the test classes can be autoloaded (which can only be the case if
- * phplist4-core is the root package, never if it is used as a library).
+ * core is the root package, never if it is used as a library).
  *
  * This is the only purpose of this interface.
  *