+++ CHANGELOG.md
@@ -7,11 +7,14 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Repository.remove (#278)
+- Interface for domain models (#274)
 - Trait for database tests and abstract web test (#264)
 - Repository methods for querying subscriptions (#253)
 - Bidirectional m:n association Subscribers/SubscriberLists (#254)
 
 ### Changed
+- Rename the Composer package to "phplist/core" (#275)
 - Remove the obsolete core classes (#267)
 - Adopt more of the default Symfony project structure (#265, #268, #269, #270)
 
@@ -20,7 +23,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
-- Remove associated subscriptions when a subscriber list or subscriber is removed (#271)
+- Remove associated subscriptions when a subscriber list or subscriber is removed (#271, #272)
 - Always truncate the DB tables after an integration test (#259)
 - Adapt the system tests to possible HTTP errors (#256)